fix for warnings
git-svn-id: http://svn.miranda-ng.org/main/trunk@13902 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Tox/src/tox_utils.cpp
@@ -65,13 +65,13 @@ bool CToxProto::IsFileExists(std::tstring path)
 	return _taccess(path.c_str(), 0) == 0;
 }
 
-MEVENT CToxProto::AddEventToDb(MCONTACT hContact, WORD type, DWORD timestamp, DWORD flags, PBYTE pBlob, DWORD cbBlob)
+MEVENT CToxProto::AddEventToDb(MCONTACT hContact, WORD type, DWORD timestamp, DWORD flags, PBYTE pBlob, size_t cbBlob)
 {
 	DBEVENTINFO dbei = { sizeof(dbei) };
 	dbei.szModule = this->m_szModuleName;
 	dbei.timestamp = timestamp;
 	dbei.eventType = type;
-	dbei.cbBlob = cbBlob;
+	dbei.cbBlob = (DWORD)cbBlob;
 	dbei.pBlob = pBlob;
 	dbei.flags = flags;
 	return db_event_add(hContact, &dbei);
